code refactoring and added test
